refactor: StageRoot UI 로직 View 클래스로 이동
